+++ src/modules/admin/adminjs.ts
@@ -1,4 +1,4 @@
-import { Webhook } from '../routes/webhook/repositories/webhook.entity';
+import { Webhook } from '../webhook/repositories/webhook.entity';
 import { AuthModule } from './auth/auth.module';
 import { AuthService } from './auth/auth.service';
