+++ src/modules/events/events.controller.ts
@@ -9,7 +9,7 @@ import { ApiTags } from '@nestjs/swagger';
 import { Observable } from 'rxjs';
 import { EventsService } from './events.service';
 import { getAddress, isAddress } from 'viem';
-import { BasicAuthGuard } from '../../auth/basic-auth.guard';
+import { BasicAuthGuard } from '../../common/guards/basic-auth.guard';
 
 @Controller('events')
 @ApiTags('events')
