+++ src/modules/health/health.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { HealthController } from './health.controller';
 import { TerminusModule } from '@nestjs/terminus';
-import { QueueModule } from '../datasources/queue/queue.module';
+import { QueueModule } from '../../datasources/queue/queue.module';
 
 @Module({
   imports: [ConfigModule, TerminusModule, QueueModule],
